#45: Re-unite package managers handlers in a unique file
